+++ packages/ricos-content/web/src/converters/draft/fromDraft/getRicosEntityData.ts
@@ -2,9 +2,22 @@ import { pickBy, identity } from 'lodash';
 /* eslint-disable fp/no-delete */
 import { TextStyle, NodeStyle } from 'ricos-schema';
 import { RicosEntityMap, RicosContentBlock } from '../../../types';
-import { LINK_PREVIEW_TYPE, EMBED_TYPE, COLLAPSIBLE_LIST_TYPE, TABLE_TYPE } from '../../../consts';
+import {
+  LINK_PREVIEW_TYPE,
+  EMBED_TYPE,
+  COLLAPSIBLE_LIST_TYPE,
+  TABLE_TYPE,
+  DIVIDER_TYPE,
+  NO_WRAP,
+} from '../../../consts';
 import { TO_RICOS_DATA_FIELD, TO_RICOS_PLUGIN_TYPE } from '../consts';
 import { convertBlockDataToRicos } from './convertRicosPluginData';
+
+const normalizeBlockData = (type, data) => {
+  if (type === DIVIDER_TYPE) {
+    data.config = data.config || { textWrap: NO_WRAP };
+  }
+};
 
 export const getEntity = (key: string | number, entityMap: RicosEntityMap) => {
   if (!entityMap[key]) {
@@ -23,6 +36,8 @@ export const getEntity = (key: string | number, entityMap: RicosEntityMap) => {
     console.error(`no data for "${blockType}"!`);
     return null;
   }
+
+  normalizeBlockData(type, data);
 
   const advancedPluginsAdditionalData = {
     [COLLAPSIBLE_LIST_TYPE]: { pairs: data?.pairs },